Dear diary,
Today I optimized the renderer to draw all actors which are currently displaying the same sprite in a batch and fixed a deadlock issue caused by the actor's internal animation timer. I also overhauled the collision detection system and put it on a separate thread. The system can now handle drawing 1000 sprites at a1ms frame interval, which is up 100x over this morning. The collision system is still pretty unstable due to excessive mutability, but I think it's time for a merge to master anyhow.

-z